+++ lib/vulcano/profile_context.rb
@@ -1,5 +1,7 @@
 # encoding: utf-8
 require 'vulcano/backend'
+require 'vulcano/rule'
+require 'vulcano/dsl'
 
 module Vulcano
   class ProfileContext
@@ -15,11 +17,54 @@ module Vulcano
       @only_ifs = only_ifs
       profile_context_owner = self
 
+      dsl = Module.new do
+        Vulcano::Resource.registry.each do |id, r|
+          define_method id.to_sym do |*args|
+            r.new(backend, *args)
+          end
+        end
+      end
+
+      rule_class = Class.new(Vulcano::Rule) do
+        include RSpec::Core::DSL
+        include dsl
+      end
+
+      outer_dsl = Class.new do
+        include dsl
+
+        define_method :rule do |*args, &block|
+          id = args[0]
+          opts = args[1] || {}
+          return if @skip_profile
+          __register_rule rule_class.new(id, opts, &block)
+        end
+
+        define_method :describe do |*args, &block|
+          path = block.source_location[0]
+          line = block.source_location[1]
+          id = "#{File.basename(path)}:#{line}"
+          rule = rule_class.new(id, {}) do
+            describe(*args, &block)
+          end
+          __register_rule rule, &block
+        end
+
+        def skip_rule(id)
+          __unregister_rule id
+        end
+
+        def only_if(&block)
+          return unless block_given?
+          @skip_profile = !block.call
+        end
+      end
+
       # This is the heart of the profile context
       # An instantiated object which has all resources registered to it
       # and exposes them to the a test file.
       # rubocop:disable Lint/NestedMethodDefinition
-      ctx = Class.new do
+      ctx = Class.new(outer_dsl) do
         include Vulcano::DSL
 
         define_method :__register_rule do |*args|
@@ -27,12 +72,6 @@ module Vulcano
         end
         define_method :__unregister_rule do |*args|
           profile_context_owner.unregister_rule(*args)
-        end
-
-        Vulcano::Resource.registry.each do |id, r|
-          define_method id.to_sym do |*args|
-            r.new(backend, *args)
-          end
         end
 
         def to_s
@@ -49,13 +88,13 @@ module Vulcano
     end
 
     def unregister_rule(id)
-      full_id = VulcanoBaseRule.full_id(@profile_id, id)
+      full_id = Vulcano::Rule.full_id(@profile_id, id)
       @rules[full_id] = nil
     end
 
     def register_rule(r)
       # get the full ID
-      full_id = VulcanoBaseRule.full_id(@profile_id, r)
+      full_id = Vulcano::Rule.full_id(@profile_id, r)
       if full_id.nil?
         # TODO: error
         return
@@ -65,7 +104,7 @@ module Vulcano
       if existing.nil?
         @rules[full_id] = r
       else
-        VulcanoBaseRule.merge(existing, r)
+        Vulcano::Rule.merge(existing, r)
       end
     end
   end
